noto: move event handling to a separate file
